requirements.txt
# Django and GIS related packages
Django==5.2
djangorestframework-gis==1.1
GDAL==3.8.4
psycopg2-binary==2.9.10

# Environment variable management
django-environ==0.12.0

# Testing tools and libraries
pytest==8.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
model-bakery==1.20.5

# Development utilities
django-extensions==4.1
